requirements.txt
PyYAML==6.0.1
openai==1.3.5
transformers==4.35.2
torch==2.1.1
scikit-learn==1.3.2
networkx==3.2.1
numpy==1.26.2
anthropic==0.5.0